修复热量机制的bug，修改热量上限强化的数值

+++ App/Data/basicData.hpp
@@ -20,7 +20,7 @@ public:
     uint8_t currentHeat{};
     uint8_t maxHeat{};
     uint8_t heatCoolDownRate{};
-    uint8_t heatCoolDownTimer{};
+    uint16_t heatCoolDownTimer{};
     uint8_t heatPerShot{};
 
 public:
